+++ R/gitrd_get_info_companies.R
@@ -1,10 +1,11 @@
 #' Reads up to date information about Bovespa companies from a github file
 #'
 #' A csv file with information about available companies and time periods is downloaded from github and read.
-#' This file is updated periodically and manually by the author.
+#' This file is updated periodically and manually by the author. When run for the first time in a R session, a .RDATA file
+#' containing the output of the function is saved in tempdir() for caching.
 #'
-#' @param type.data A string that sets the type of information to be returned ('companies' or 'companies&files').
-#' If 'companies', it will return a dataframe with several information about companies, but without files download link.
+#' @param type.data A string that sets the type of information to be returned ('companies' or 'companies_files').
+#' If 'companies', it will return a dataframe with several information about companies, but without download links.
 #'
 #' @return A dataframe with several information about Bovespa companies
 #' @export
@@ -15,13 +16,21 @@
 #' df.info <- gitrd.get.info.companies()
 #' str(df.info)
 #' }
-gitrd.get.info.companies <- function(type.data = 'companies&files') {
+gitrd.get.info.companies <- function(type.data = 'companies_files') {
 
   # error checking
-  possible.values <- c('companies&files', 'companies')
-
+  possible.values <- c('companies_files', 'companies')
   if ( !(type.data %in% possible.values) ) {
     stop('Input type.data should be one of:\n\n', paste0(possible.values, collapse = '\n'))
+  }
+
+  # check if cache file exists
+  my.f.rdata <- file.path(tempdir(),paste0('df_info_', type.data, '.RData') )
+
+  if (file.exists(my.f.rdata)) {
+    cat('Found cache file. Loading data..')
+    load(my.f.rdata)
+    return(df.info)
   }
 
   # get data from github
@@ -61,7 +70,9 @@ gitrd.get.info.companies <- function(type.data = 'companies&files') {
 
   if (type.data == 'companies') {
 
-    df.info.agg <- unique(df.info[, -c(7:10)])
+    my.cols <- my.cols <- c("name.company","id.company", "situation",
+                            "main.sector", "sub.sector", "segment", "tickers")
+    df.info.agg <- unique(df.info[, my.cols])
 
     my.fun <- function(df) {
       return(c(min(df$id.date), max(df$id.date)))
@@ -77,11 +88,12 @@ gitrd.get.info.companies <- function(type.data = 'companies&files') {
     df.info.agg$first.date <- as.Date(df.info.agg$first.date)
     df.info.agg$last.date <- as.Date(df.info.agg$last.date)
 
-    return(df.info.agg)
-
-  } else {
-    return(df.info)
+    df.info <- df.info.agg
   }
 
+  cat('\nCaching RDATA into tempdir()')
+  save('df.info', file = my.f.rdata)
+
+  return(df.info)
 
 }