+++ NEWS.md
@@ -1,7 +1,12 @@
-### Version 0.X (2017-XX-XX)
+### Version 0.6 (2017-10-21)
 
-- [x] fixed bug with NA dates and names of companies
-- [ ] defaults with dollar adjusted
+- fixed bug with NA dates and names of companies
+- Implemented inflation adjusted values
+- Added current capital composition
+- added tickers list for each company
+- Dividends history is now available
+- Yearly data (DFP system) is now also available
+- Users can now select max levels in account description (rows)
 
 ### Version 0.5 (2017-09-29)